requirements.txt
streamlit
crewai
pysqlite3-binary
python-dotenv
langchain-community
chromadb
langchain-google-genai
